+++ backend/src/main/java/org/codecritique/thrifty/entity/BaseEntity.java
@@ -15,9 +15,9 @@ public abstract class BaseEntity implements Serializable {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    protected Integer id;
+    protected Long id;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
